Fixes #26078 - cleanup window call, calls go through redux.
