+++ app/assets/javascripts/host_edit_interfaces.js
@@ -3,20 +3,6 @@ $(document).ready(function() {
   $('#host_name').focus();
 })
 
-function remove_interface(interface_id) {
-  tfm.hosts.removeInterface(interface_id);
-}
-
-function add_interface() {
-  const form = get_interface_template_clone();
-  show_interface_modal(form);
-}
-
-function edit_interface(interface_id) {
-  const form = $('#interfaces #interfaceHidden'+interface_id).clone(true);
-  show_interface_modal(form);
-}
-
 function show_interface_modal(modal_content) {
   var modal_window = $('#interfaceModal');
   var interface_id = modal_content.data('interface-id');
@@ -24,7 +10,6 @@ function show_interface_modal(modal_content) {
   modal_window.data('current-id', interface_id);
 
   var identifier = modal_content.find('.interface_identifier').val();
-
 
   modal_window.find('.modal-body').html('');
   modal_window.find('.modal-body').append(modal_content.contents());
@@ -38,6 +23,7 @@ function show_interface_modal(modal_content) {
 function save_interface_modal() {
   var modal_window = $('#interfaceModal');
   var interface_id = modal_window.data('current-id');
+  var is_new = modal_window.data('new-interface');
 
   //destroy ui tools so when opening the modal again they will show correctly
   modal_window.find('a[rel="popover-modal"]').popover('destroy');
@@ -48,12 +34,16 @@ function save_interface_modal() {
 
   var modal_form = modal_window.find('.modal-body').contents();
 
-  var interface_hidden = get_interface_hidden(interface_id);
+  var interface_hidden = $(tfm.hosts.getInterfaceHiddenFields(interface_id));
   interface_hidden.html('');
   interface_hidden.append(modal_form);
 
   close_interface_modal();
-  update_interface_info(interface_id, interface_hidden);
+  if (is_new) {
+    add_interface(interface_id, interface_hidden);
+  } else {
+    update_interface_info(interface_id, interface_hidden);
+  }
 }
 
 function close_interface_modal() {
@@ -61,45 +51,12 @@ function close_interface_modal() {
 
   modal_window.modal('hide');
   modal_window.removeData('current-id');
+  modal_window.removeData('new-interface');
   modal_window.find('.modal-body').html('');
 }
 
-function get_interface_template_clone() {
-  var content = $('#interfaces .interfaces_fields_template').html();
-  var interface_id = new Date().getTime();
-
-  content = fix_template_names(content, 'interfaces', interface_id);
-
-  var hidden = $(content);
-
-  $('#interfaceForms').closest("form").trigger({type: 'nested:fieldAdded', field: hidden});
-  $('a[rel="popover"]').popover();
-  $('a[rel="twipsy"]').tooltip();
-
-  hidden.attr('id', 'interfaceHidden'+interface_id);
-  hidden.data('interface-id', interface_id);
-  hidden.find('.destroyFlag').val(0);
-
-  return hidden;
-}
-
-function get_interface_hidden(interface_id) {
-  var interface_hidden = $('#interfaceHidden'+interface_id);
-  if ( interface_hidden.length == 0) {
-
-    interface_hidden = $('<div></div>');
-    interface_hidden.attr('class', 'hidden');
-    interface_hidden.attr('id', 'interfaceHidden'+interface_id);
-    interface_hidden.data('interface-id', interface_id);
-    interface_hidden.data('new-interface', true);
-
-    $('#interfaceForms').append(interface_hidden);
-  }
-  return interface_hidden;
-}
-
-function update_interface_info(interface_id, interface_form) {
-  const interfaceData = {
+function get_interface_data(interface_form) {
+  return {
     name: interface_form.find('.interface_name').val(),
     domain: interface_form.find('.interface_domain option:selected').text(),
     type: interface_form.find('.interface_type').val(),
@@ -112,50 +69,22 @@ function update_interface_info(interface_id, interface_form) {
     provision: interface_form.find('.interface_provision').is(':checked'),
     virtual: interface_form.find('.virtual').is(':checked'),
     attachedTo: interface_form.find('.attached').val(),
+    providerSpecificInfo: providerSpecificInterfaceInfo(interface_form),
     hasErrors: interface_form.find('.has-error').length > 0,
+    editing: false,
   };
-  if (interface_form.data('new-interface')) {
-    tfm.hosts.addInterface({id: interface_id, ...interfaceData});
-    interface_form.removeData('new-interface');
-  } else {
-    tfm.hosts.updateInterface(interface_id, interfaceData);
-  }
 }
 
-function updateInterfaceTable(state, prevState) {
-  var pathname = window.location.pathname;
-
-  state.interfaces.forEach(function(interfaceData, idx) {
-    var prevData = prevState.interfaces.filter(function(i){ return i.id == interfaceData.id })[0];
-    if (prevData && prevData == interfaceData) return;
-
-    var hiddenFields = get_interface_hidden(interfaceData.id)[0];
-    var nameEl = hiddenFields.getElementsByClassName('interface_name')[0];
-    var primaryCheck = hiddenFields.getElementsByClassName('interface_primary')[0];
-    var provisionCheck = hiddenFields.getElementsByClassName('interface_provision')[0];
-    var virtualCheck = hiddenFields.getElementsByClassName('virtual')[0];
-
-    nameEl.value = interfaceData.name;
-    primaryCheck.checked = interfaceData.primary;
-    provisionCheck.checked = interfaceData.provision;
-    virtualCheck.checked = interfaceData.virtual;
-
-    if (interfaceData.primary) {
-      $('#host_name').val(interfaceData.name);
-
-      var fqdn_val = tfm.hosts.fqdn(interfaceData.name, interfaceData.domain);
-      if (fqdn_val.length > 0 && pathname === '/hosts/new') {
-        tfm.breadcrumbs.updateTitle(__("Create Host") + " | " + fqdn_val);
-      }
-    }
-  });
-  state.destroyed.forEach(function(destroyedId, idx) {
-    $('#interfaceHidden'+destroyedId+' .destroyFlag').val(1);
-  });
+function add_interface(interface_id, interface_form) {
+  var interfaceData = get_interface_data(interface_form);
+  interfaceData.id = interface_id;
+  tfm.hosts.addInterface(interfaceData);
 }
-tfm.observeStore(updateInterfaceTable, function(store) {
-  return store.hosts.interfaces;
-});
+
+function update_interface_info(interface_id, interface_form) {
+  var interfaceData = get_interface_data(interface_form);
+  tfm.hosts.updateInterface(interface_id, interfaceData);
+}
 
 function active_interface_forms() {
   return $.grep($('#interfaceForms > div'), function(f) {
@@ -212,6 +141,11 @@ $(document).on('change', '#host_name', function () {
 });
 
 var providerSpecificNICInfo = null;
+function providerSpecificInterfaceInfo(formFields) {
+  if (typeof window.providerSpecificNICInfo === 'function')
+    return window.providerSpecificNICInfo($(formFields));
+  return null;
+};
 
 $(document).on('change', '.compute_attributes', function () {
   // remove "from profile" note if any of the fields changes
