+++ app/views/hosts/_interfaces.html.erb
@@ -8,7 +8,7 @@
   <span id="interfaceListWrapper"></span>
 
   <div class="btn-toolbar">
-    <button type="button" onclick="add_interface();return false;"
+    <button type="button" onclick="tfm.hosts.showNewInterfaceModal();return false;"
       class="info btn btn-success" id="addInterface"
                                    title="add new network interface">+ <%= _('Add Interface') %>
     </button>
