+++ webpack/assets/javascripts/foreman_nested_forms.js
@@ -1,0 +1,4 @@
+export function fixTemplateNames(content, assoc, newId) {
+  const regexp = new RegExp(`new_${assoc}`, 'g');
+  return content.replace(regexp, newId);
+}
