+++ webpack/assets/javascripts/react_app/redux/actions/hosts/interfaces.js
@@ -3,6 +3,7 @@ import {
   INTERFACES_ADD_INTERFACE,
   INTERFACES_UPDATE_INTERFACE,
   INTERFACES_REMOVE_INTERFACE,
+  INTERFACES_TOGGLE_INTERFACE_EDITING,
   INTERFACES_SET_PRIMARY_INTERFACE_NAME,
   INTERFACES_SET_PRIMARY_INTERFACE,
   INTERFACES_SET_PROVISION_INTERFACE,
@@ -29,6 +30,11 @@ export const updateInterface = (id, newValues) => ({
 export const removeInterface = id => ({
   type: INTERFACES_REMOVE_INTERFACE,
   payload: { id },
+});
+
+export const toggleInterfaceEditing = (id, flag) => ({
+  type: INTERFACES_TOGGLE_INTERFACE_EDITING,
+  payload: { id, flag },
 });
 
 export const setPrimaryInterface = id => ({
