+++ webpack/assets/javascripts/react_app/redux/consts.js
@@ -35,6 +35,8 @@ export const INTERFACES_INITIALIZE = 'INTERFACES_INITIALIZE';
 export const INTERFACES_ADD_INTERFACE = 'INTERFACES_ADD_INTERFACE';
 export const INTERFACES_UPDATE_INTERFACE = 'INTERFACES_UPDATE_INTERFACE';
 export const INTERFACES_REMOVE_INTERFACE = 'INTERFACES_REMOVE_INTERFACE';
+export const INTERFACES_TOGGLE_INTERFACE_EDITING =
+  'INTERFACES_TOGGLE_INTERFACE_EDITING';
 export const INTERFACES_SET_PRIMARY_INTERFACE =
   'INTERFACES_SET_PRIMARY_INTERFACE';
 export const INTERFACES_SET_PROVISION_INTERFACE =
